Moved Prov parser to seperate util with String input #4343
